+++ src/AppFactory.Framework.Logging.Abstractions/ILoggerFactory.cs
@@ -1,0 +1,9 @@
+﻿using Serilog;
+
+namespace AppFactory.Framework.Logging;
+
+public interface ILoggerFactory
+{
+    ILogger CreateLogger(LoggerConfiguration configuration);
+    ILogger CreatePlainTextLogger();
+}